Fix non-existant permission check. fixes #673.
git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@2751 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/acl.php
@@ -208,7 +208,7 @@ class ACL {
 			return $name;
 		}
 		$name = self::normalize_permission( $name );
-		return (int) DB::get_value( 'SELECT id FROM {tokens} WHERE name=?', array( $name ) );
+		return DB::get_value( 'SELECT id FROM {tokens} WHERE name=?', array( $name ) );
 	}
 
 	/**
@@ -285,7 +285,7 @@ class ACL {
 		 * Do we allow perms that don't exist?
 		 * When ACL is functional ACCESS_NONEXISTANT_PERMISSION should be false by default.
 		 */
-		if ( is_null( $permission ) ) {
+		if ( is_null( $permission) ) {
 			return self::ACCESS_NONEXISTANT_PERMISSION;
 		}
